+++ src/QuartoNotebookRunner.jl
@@ -43,6 +43,10 @@ export Server, render, run!, close!
 
 # Includes.
 
+const QNR_VERSION =
+    VersionNumber(TOML.parsefile(joinpath(@__DIR__, "..", "Project.toml"))["version"])
+include_dependency(joinpath(@__DIR__, "..", "Project.toml"))
+
 include("UserError.jl")
 include("Malt.jl")
 include("WorkerSetup.jl")